pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dosboxlauncher"
version = "0.1.0"
description = "A GUI for managing and launching DOS games with DOSBox"
readme = "Readme.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "DOSBoxLauncher"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Emulators",
]

dependencies = [
    "PyGObject>=3.44.0,<3.50.0",
    "pycairo>=1.27.0",
    "PyYAML>=6.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
dosboxlauncher = "dosboxlauncher.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E402"]
exclude = ["tooltips.py"]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
